the seach results bar on the homepage now renders appropriately
